Move the Dialog enum to its own file
