+++ src/main/java/sam/parser/SamInvalidDateException.java
@@ -1,13 +1,13 @@
 package sam.parser;
 
+import sam.Dialog;
 import sam.SamException;
-import sam.Ui;
 
 /**
  * A SamException thrown when an error occurs during parsing a date.
  */
 public class SamInvalidDateException extends SamException {
     public SamInvalidDateException() {
-        super(Ui.Dialog.INVALID_DATE.getDialog());
+        super(Dialog.INVALID_DATE.getDialog());
     }
 }
