+++ src/main/java/sam/parser/SamInvalidIntException.java
@@ -1,12 +1,12 @@
 package sam.parser;
 
+import sam.Dialog;
 import sam.SamException;
-import sam.Ui;
 /**
  * A SamException thrown when an error occurs during parsing an integer.
  */
 public class SamInvalidIntException extends SamException {
     public SamInvalidIntException() {
-        super(Ui.Dialog.INVALID_INT.getDialog());
+        super(Dialog.INVALID_INT.getDialog());
     }
 }
