+++ src/main/java/sam/parser/SamInvalidTaskException.java
@@ -1,13 +1,13 @@
 package sam.parser;
 
+import sam.Dialog;
 import sam.SamException;
-import sam.Ui;
 
 /**
  * A SamException thrown when a non-existing task is retrieved.
  */
 public class SamInvalidTaskException extends SamException {
     public SamInvalidTaskException() {
-        super(Ui.Dialog.INVALID_TASK.getDialog());
+        super(Dialog.INVALID_TASK.getDialog());
     }
 }
