+++ src/main/java/sam/parser/SamUnknownCommandException.java
@@ -1,13 +1,13 @@
 package sam.parser;
 
+import sam.Dialog;
 import sam.SamException;
-import sam.Ui;
 
 /**
  * A SamException thrown when the user gives an unknown command.
  */
 public class SamUnknownCommandException extends SamException {
     public SamUnknownCommandException() {
-        super(Ui.Dialog.UNKNOWN_COMMAND.getDialog());
+        super(Dialog.UNKNOWN_COMMAND.getDialog());
     }
 }
