+++ src/main/java/sam/storage/SamSaveFailedException.java
@@ -1,13 +1,13 @@
 package sam.storage;
 
+import sam.Dialog;
 import sam.SamException;
-import sam.Ui;
 
 /**
  * A SamException thrown when an error occurs during saving the task list.
  */
 public class SamSaveFailedException extends SamException {
     public SamSaveFailedException() {
-        super(Ui.Dialog.SAVE_FAILED.getDialog());
+        super(Dialog.SAVE_FAILED.getDialog());
     }
 }
