+++ src/main/java/sam/task/SamMissingTaskArgException.java
@@ -1,7 +1,7 @@
 package sam.task;
 
+import sam.Dialog;
 import sam.SamException;
-import sam.Ui;
 
 /**
  * A SamException thrown when a required task argument
@@ -9,6 +9,6 @@ import sam.Ui;
  */
 public class SamMissingTaskArgException extends SamException {
     public SamMissingTaskArgException() {
-        super(Ui.Dialog.MISSING_TASK_ARG.getDialog());
+        super(Dialog.MISSING_TASK_ARG.getDialog());
     }
 }
