+++ src/main/java/sam/task/SamMissingTaskException.java
@@ -1,7 +1,7 @@
 package sam.task;
 
+import sam.Dialog;
 import sam.SamException;
-import sam.Ui;
 
 /**
  * A SamException thrown when a required task id
@@ -9,6 +9,6 @@ import sam.Ui;
  */
 public class SamMissingTaskException extends SamException {
     public SamMissingTaskException() {
-        super(Ui.Dialog.MISSING_TASK.getDialog());
+        super(Dialog.MISSING_TASK.getDialog());
     }
 }
